Ajoute page ajouter produit Ajoute refus sur devis Ajout process transformation devis en commande

+++ database/migrations/2015_11_15_145631_create_products_table.php
@@ -18,7 +18,7 @@ class CreateProductsTable extends Migration
             $table->string('description');
             $table->enum('type',['time','one_shot']);
             $table->smallInteger('value')->unsigned()->default(null)->nullable();
-            $table->smallInteger('price')->unsigned();
+            $table->decimal('price')->unsigned();
         });
     }
